pyproject.toml
[project]
name = "skillproposer"
version = "0.1.0"
description = "SkillProposer project"
requires-python = ">=3.10"
license = "MIT"
dependencies = [
    "django",
    "django-environ",
    "chromadb",
    "langchain",
    "langchain-community",
    "langchain-text-splitters",
    "langgraph",
    "langchain-core",
    "google-generativeai",
    "unstructured[docx,pdf,md]",
    "python-magic-bin",
    "python-dotenv",
    "requests",
    "sentence-transformers",
    "pydantic",
    "langchain-google-genai",
    "langchain-openai",
]

[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
packages = []
py-modules = []
